randomize wait times, support permanent and session ban for parents

+++ src/database/sqlite/sqlite_aclk.c
@@ -116,7 +116,7 @@ static int create_host_callback(void *data, int argc, char **argv, char **column
     struct rrdhost_system_info *system_info = callocz(1, sizeof(struct rrdhost_system_info));
     __atomic_sub_fetch(&netdata_buffers_statistics.rrdhost_allocations_size, sizeof(struct rrdhost_system_info), __ATOMIC_RELAXED);
 
-    system_info->hops = str2i((const char *) argv[IDX_HOPS]);
+    system_info->hops = (int16_t)str2i((const char *) argv[IDX_HOPS]);
 
     sql_build_host_system_info((nd_uuid_t *)argv[IDX_HOST_ID], system_info);
